requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.6.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==40.0.2
dataclasses-json==0.5.7
Deprecated==1.2.13
dnspython==2.3.0
et-xmlfile==1.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
joblib==1.2.0
langchain==0.0.162
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
pinecone-client==2.2.1
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pypandoc==1.11
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.6.4
urllib3==2.0.2
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.9.2
